// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ReservationStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled'
} as const

export type ReservationStatus = (typeof ReservationStatus)[keyof typeof ReservationStatus]


export const NotificationType = {
  reservation_cancelled: 'reservation_cancelled'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]